requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.1
attrs==19.1.0
autopep8==1.5.2
black==19.3b0
Click==7.0
distlib==0.3.0
Django==3.0.6
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pylint==2.4.2
pytz==2020.1
six==1.12.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.0
virtualenv==20.0.18
wrapt==1.11.2
zipp==3.1.0
